refactor : Golbãlobal í패키지 구조 변경ã (Global 패키지 생성)
