+++ src/main/java/com/example/block/global/apiPayload/code/BaseCode.java
@@ -1,4 +1,4 @@
-package com.example.block.apiPayload.code;
+package com.example.block.global.apiPayload.code;
 
 public interface BaseCode {
     public ReasonDTO getReason();
