+++ src/main/java/com/example/block/global/apiPayload/code/BaseErrorCode.java
@@ -1,4 +1,4 @@
-package com.example.block.apiPayload.code;
+package com.example.block.global.apiPayload.code;
 
 public interface BaseErrorCode {
     public ErrorReasonDTO getReason();
