+++ src/main/java/com/example/block/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.example.block.config;
+package com.example.block.global.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.models.Components;
